+++ web-registry/src/ledger/api/queries/ecocredit.ts
@@ -1,0 +1,226 @@
+import {
+  QueryClientImpl,
+  DeepPartial,
+  QueryBalanceRequest,
+  QueryBalanceResponse,
+  QueryBatchesRequest,
+  QueryBatchesResponse,
+  QueryBatchInfoRequest,
+  QueryBatchInfoResponse,
+  QueryClassesRequest,
+  QueryClassesResponse,
+  QueryClassInfoRequest,
+  QueryClassInfoResponse,
+  QueryCreditTypesRequest,
+  QueryCreditTypesResponse,
+} from '@regen-network/api/lib/generated/regen/ecocredit/v1alpha1/query';
+
+// TODO: pagination not implemented yet
+
+/**
+ *
+ * ECOCREDIT MODULE QUERIES
+ * ---------------------
+ *  - Balance
+ *  - BatchInfo
+ *  - Batches
+ *  - ClassInfo
+ *  - Classes
+ *  - CreditTypes
+ *  - Params        (TODO)
+ *  - Supply        (TODO)
+ *
+ */
+
+/**
+ *
+ * QUERY TYPES
+ *
+ */
+
+// typing the query client
+
+export type EcocreditQueryClient = QueryClientImpl;
+
+interface EcocreditQueryClientProps {
+  client: EcocreditQueryClient;
+}
+
+// typing and linking query names and corresponding input params
+
+type BalanceParams = {
+  query: 'balance';
+  params: DeepPartial<QueryBalanceRequest>;
+};
+
+type BatchInfoParams = {
+  query: 'batchInfo';
+  params: DeepPartial<QueryBatchInfoRequest>;
+};
+
+type BatchesParams = {
+  query: 'batches';
+  params: DeepPartial<QueryBatchesRequest>;
+};
+
+type ClassInfoParams = {
+  query: 'classInfo';
+  params: DeepPartial<QueryClassInfoRequest>;
+};
+
+type ClassesParams = {
+  query: 'classes';
+  params: DeepPartial<QueryClassesRequest>;
+};
+
+type CreditTypesParams = {
+  query: 'creditTypes';
+  params: DeepPartial<QueryCreditTypesRequest>;
+};
+
+export type EcocreditQueryProps =
+  | BalanceParams
+  | BatchInfoParams
+  | BatchesParams
+  | ClassInfoParams
+  | ClassesParams
+  | CreditTypesParams;
+
+// typing the response
+
+export type EcocreditQueryResponse =
+  | QueryBalanceResponse
+  | QueryBatchInfoResponse
+  | QueryBatchesResponse
+  | QueryClassInfoResponse
+  | QueryClassesResponse
+  | QueryCreditTypesResponse;
+
+/**
+ *
+ * QUERY FUNCTIONS
+ *
+ */
+
+// Balance
+
+interface QueryBalanceProps extends EcocreditQueryClientProps {
+  request: DeepPartial<QueryBalanceRequest>;
+}
+
+export const queryBalance = async ({
+  client,
+  request,
+}: QueryBalanceProps): Promise<QueryBalanceResponse> => {
+  try {
+    return await client.Balance({
+      account: request.account,
+      batchDenom: request.batchDenom,
+    });
+  } catch (err) {
+    throw new Error(
+      `Error in the Balance query of the ledger ecocredit module: ${err}`,
+    );
+  }
+};
+
+// Batch info
+
+interface QueryBatchInfoProps extends EcocreditQueryClientProps {
+  request: DeepPartial<QueryBatchInfoRequest>;
+}
+
+export const queryBatchInfo = async ({
+  client,
+  request,
+}: QueryBatchInfoProps): Promise<QueryBatchInfoResponse> => {
+  try {
+    return await client.BatchInfo({
+      batchDenom: request.batchDenom,
+    });
+  } catch (err) {
+    throw new Error(
+      `Error in the BatchInfo query of the ledger ecocredit module: ${err}`,
+    );
+  }
+};
+
+// Batches
+
+interface QueryBatchesProps extends EcocreditQueryClientProps {
+  request: DeepPartial<QueryBatchesRequest>;
+}
+
+export const queryBatches = async ({
+  client,
+  request,
+}: QueryBatchesProps): Promise<QueryBatchesResponse> => {
+  try {
+    return await client.Batches({
+      classId: request.classId,
+    });
+  } catch (err) {
+    throw new Error(
+      `Error in the Batches query of the ledger ecocredit module: ${err}`,
+    );
+  }
+};
+
+// Class info
+
+interface QueryClassInfoProps extends EcocreditQueryClientProps {
+  request: DeepPartial<QueryClassInfoRequest>;
+}
+
+export const queryClassInfo = async ({
+  client,
+  request,
+}: QueryClassInfoProps): Promise<QueryClassInfoResponse> => {
+  try {
+    return await client.ClassInfo({
+      classId: request.classId,
+    });
+  } catch (err) {
+    throw new Error(
+      `Error in the ClassInfo query of the ledger ecocredit module: ${err}`,
+    );
+  }
+};
+
+// Classes
+
+interface QueryClassesProps extends EcocreditQueryClientProps {
+  request: DeepPartial<QueryClassesRequest>;
+}
+
+export const queryClasses = async ({
+  client,
+  request,
+}: QueryClassesProps): Promise<QueryClassesResponse> => {
+  try {
+    return await client.Classes({});
+  } catch (err) {
+    throw new Error(
+      `Error in the Classes query of the ledger ecocredit module: ${err}`,
+    );
+  }
+};
+
+// CreditTypes
+
+interface QueryCreditTypesProps extends EcocreditQueryClientProps {
+  request: DeepPartial<QueryCreditTypesRequest>;
+}
+
+export const queryCreditTypes = async ({
+  client,
+  request,
+}: QueryCreditTypesProps): Promise<QueryCreditTypesResponse> => {
+  try {
+    return await client.CreditTypes({});
+  } catch (err) {
+    throw new Error(
+      `Error in the CreditTypes query of the ledger ecocredit module: ${err}`,
+    );
+  }
+};
