+++ web-registry/src/ledger/hooks/useMsgClient.ts
@@ -1,8 +1,8 @@
 import { useState, useCallback } from 'react';
 import { StdFee, DeliverTxResponse } from '@cosmjs/stargate';
 
-import { useLedger } from '../ledger';
-import { Wallet } from '../lib/wallet';
+import { useLedger } from '../context/LedgerContext';
+import { Wallet } from '../../lib/wallet';
 
 interface TxData {
   msgs: any[];
