+++ web-registry/src/pages/CreditClassDetails.tsx
@@ -11,7 +11,7 @@ import { CreditClassDetailsWithContent } from './CreditClassDetailsWithContent';
 import { CreditClassDetailsSimple } from './CreditClassDetailsSimple';
 import { queryEcoClassInfo } from '../lib/ecocredit';
 import { getMetadata } from '../lib/metadata-graph';
-import { onChainClassRegExp } from '../lib/ledger';
+import { onChainClassRegExp } from '../ledger';
 import { ClassInfo } from '../types/ledger/ecocredit';
 
 interface CreditDetailsProps {
